requirements.txt
# 核心依赖
python-dotenv==1.0.0

# HTTP请求
requests==2.31.0

# PDF处理
PyMuPDF==1.23.8
pdfplumber==0.10.3
PyPDF2==3.0.1

# OCR和表格识别
pytesseract==0.3.10
camelot-py[cv]==0.11.0
tabula-py==2.9.0

# LLM API
openai==1.12.0
anthropic==0.18.0

# 数据处理
beautifulsoup4==4.12.3
lxml==5.1.0

# 进度条
tqdm==4.66.1

# 日志和配置
pyyaml==6.0.1

# arXiv API (可选，用于直接搜索arXiv)
arxiv==2.1.0

